[#741] refactor switch from event API to event Transformers
