+++ src/features/Calendars/services/updateSeriesAsync.ts
@@ -1,9 +1,10 @@
-import { updateSeries } from '@/features/Events/api/updateSeries'
+import { makeSeriesJCal } from '@/features/Events/transformers/makeSeriesJCal'
 import { CalendarEvent } from '@/features/Events/EventsTypes'
 import { toRejectedError } from '@/utils/errorUtils'
 import { createAsyncThunk } from '@reduxjs/toolkit'
 import { Calendar } from '../CalendarTypes'
 import { RejectedError } from '../types/RejectedError'
+import { fetchAllRecurrentVevents, putEvent } from '@/features/Events/EventDao'
 
 export const updateSeriesAsync = createAsyncThunk<
   void,
@@ -13,7 +14,15 @@ export const updateSeriesAsync = createAsyncThunk<
   'calendars/updateSeries',
   async ({ cal, event, removeOverrides = true }, { rejectWithValue }) => {
     try {
-      await updateSeries(event, cal.owner?.emails?.[0], removeOverrides)
+      const vevents = await fetchAllRecurrentVevents(event)
+      const jCal = makeSeriesJCal(
+        vevents,
+        event,
+        cal.owner?.emails?.[0],
+        removeOverrides
+      )
+
+      await putEvent(event, jCal)
     } catch (err) {
       return rejectWithValue(toRejectedError(err))
     }
