--- src/features/Events/EventApi.ts
@@ -1,6 +1,0 @@
-import { extractEventBaseUuid } from '@/utils/extractEventBaseUuid'
-import { TIMEZONES } from '@/utils/timezone-data'
-import ICAL from 'ical.js'
-import { CounterProposalPayload, postCounterProposalRaw } from './EventDao'
-import { CalendarEvent } from './EventsTypes'
-import { makeTimezone, makeVevent } from './utils'
